requirements.txt
django
djangorestframework
pandas
numpy
transformers
nltk
spacy
requests
pymysql
# psycopg2
mysql-connector-python
python-dotenv
